fix 퀴즈 객관식문제페이지 컴포넌트화
